pyproject.toml
[tool.poetry]
name = "text_to_audiobook_app"
version = "0.1.0"
description = ""
authors = ["Larry Hudson <larryhudson@hey.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
requests = "^2.26.0"
azure-cognitiveservices-speech = "^1.19.0"
azure-cognitiveservices-vision-computervision = "^0.9.0"
Pillow = "^8.4.0"
textract = "^1.6.4"
trafilatura = "^1.0.0"
python-decouple = "^3.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
